requirements.txt
Django==4.2.13
Pillow==10.0.0
pandas==2.0.3
openpyxl==3.1.2
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
gunicorn==21.2.0
python-decouple==3.8
whitenoise==6.5.0
